Refactor MySQL connection initialization in plugins
Reworked the initialization of the MySQL connections in LoungeSystemPlugin, GreeterPlugin, and RankSystemPlugin. Rather than using a static readonly helper, connections are now created and managed within the plugin's OnLoad method. It includes table initialization and closing of connections within the method. Additionally, added a reference to the CommonPluginHelpers project in the NinjaBot-DC project.