Add DefaultPlugin class, update package versions and plugin/xml files
A DefaultPlugin class is added to reduce redundancy in code. Packages in NinjaBot-DC, CommonPluginHelpers and PluginBase have been updated to the latest version. plugin.xml files are also added for better metadata handling of plugins, and relevant changes are done in .cs files to accommodate this. PostBuild commands are also updated to copy these XML files to the appropriate directories.