Refactor LoungeTrustUserButton and clean code
The code for LoungeTrustUserButton has been refactored, replacing the dropdown user selection with a DiscordUserSelectComponent. Unused using directives and lines of code in several areas have been removed for code cleanliness. Additionally, the original interaction response is now deleted after the permissions are updated in both LoungeTrustUserButton and LoungeUnTrustUserButton actions.