Implement cancellation token in plugins
A cancellation token has been added to the base plugin and the rank system plugin to support task cancellation. A cancellation token source is created and linked to the bot's cancellation token, allowing concurrent tasks to be cancelled if the bot is stopped. An instance of the cancellation token source is also introduced in the main worker class.