Add NewLoungeHelper and refactor VoiceStateUpdated event
A new helper class named NewLoungeHelper has been created and utilized in the LoungeSystemPlugin, aimed to improve code maintainability. The creation of new lounges, which was initially handled in the VoiceStateUpdated event, has now been refactored and moved to this helper. The VoiceStateUpdated event now simply calls a method from the helper, de-cluttering its codebase and enhancing readability.